pyproject.toml
[tool.poetry]
name = "audit-trail-fastapi"
version = "0.1.0"
description = ""
authors = ["Bogdan Chivescu <bogdan.chivescu@clarisoft.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.85.2"
uvicorn = "^0.19.0"
python-dotenv ="0.21.0"
pymongo="4.3.2"
fastapi-utils="0.2.1"

[tool.poetry.dev-dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
